requirements.txt
streamlit==1.11.0
requests==2.25.1
urllib3==1.26.5
pyrebase4==4.3.0
six==1.15.0
pandas==1.3.3
python-dateutil==2.8.2
pymongo